# docs/source/spec/openapi.yml
openapi: 3.0.2
info:
  title: Model Zoo
  description: ""
  version: 1.0.10
paths:
...
  "/experiments/{exp_id}":
    get:
      tags:
      - Experiments
      summary: Get Experiment
      description: |-
        Get experiment by ID.

        Args:
            exp_id (str): Experiment ID
            connector (Connector): Connector to use
            return_plots (bool, optional): If plots should be returned. Defaults to True.
            return_artifacts (bool, optional): If artifacts should be returned. Defaults to True.

        Raises:
            HTTPException: 404 Not Found if experiment does not exist
            HTTPException: 500 Internal Server Error if error occurs

        Returns:
            Dict: Experiment details
      operationId: get_experiment_experiments__exp_id__get
      parameters:
      - required: true
        schema:
          title: Exp Id
          type: string
        name: exp_id
        in: path
      - required: true
        schema:
          "$ref": "#/components/schemas/src__models__experiment__Connector"
        name: connector
        in: query
      - required: false
        schema:
          title: Return Plots
          type: boolean
          default: true
        name: return_plots
        in: query
      - required: false
        schema:
          title: Return Artifacts
          type: boolean
          default: true
        name: return_artifacts
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExperimentResponse"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearerWithCookie: []
  "/experiments/clone":
    post:
      tags:
      - Experiments
      summary: Clone Experiment
      description: |-
        Clone experiment. (Not implemented yet.)

        Args:
            item (ClonePackageModel): Item to clone
            connector (Connector): Experiment connector

        Returns:
            Dict: Details of cloned experiment
      operationId: clone_experiment_experiments_clone_post
      parameters:
      - required: true
        schema:
          "$ref": "#/components/schemas/src__models__experiment__Connector"
        name: connector
        in: query
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ClonePackageModel"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearerWithCookie: []
...
components:
  schemas:
    Artifact:
      title: Artifact
      required:
      - artifactType
      - name
      - url
      type: object
      properties:
        artifactType:
          title: Artifacttype
          type: string
        name:
          title: Name
          type: string
        url:
          title: Url
          type: string
        timestamp:
          title: Timestamp
          type: string
        framework:
          title: Framework
          type: string
      description: |-
        Artifact model for model cards.
        Artifacts can be linked models, datasets, or other artifacts.
...
    ClonePackageModel:
      title: ClonePackageModel
      required:
      - id
      type: object
      properties:
        id:
          title: Id
          type: string
        clone_name:
          title: Clone Name
          type: string
      description: Model for cloning a experiment. Currently unused.
...
    ExperimentResponse:
      title: ExperimentResponse
      required:
      - id
      - owner
      - name
      - project_name
      - tags
      - frameworks
      - config
      type: object
      properties:
        id:
          title: Id
          type: string
        owner:
          title: Owner
          type: string
        name:
          title: Name
          type: string
        project_name:
          title: Project Name
          type: string
        tags:
          title: Tags
          type: array
          items:
            type: string
        frameworks:
          title: Frameworks
          type: array
          items:
            type: string
        config:
          title: Config
          type: object
        scalars:
          title: Scalars
          type: array
          items:
            type: object
        plots:
          title: Plots
          type: array
          items:
            type: object
        artifacts:
          title: Artifacts
          type: object
          additionalProperties:
            "$ref": "#/components/schemas/Artifact"
      description: Response model for getting an experiment.
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            "$ref": "#/components/schemas/ValidationError"
...
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
            - type: string
            - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
    src__models__experiment__Connector:
      title: Connector
      enum:
      - ''
      - clearml
      type: string
      description: Allowed connectors for experiments.
  securitySchemes:
    OAuth2PasswordBearerWithCookie:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: "/auth/"
...